Extracted a separate class - parser for Uwazi forms WIP
