+++ mobile/src/main/java/rs/readahead/washington/mobile/views/fragment/uwazi/entry/UwaziEntryFragment.kt
@@ -12,24 +12,18 @@ import androidx.core.app.ActivityCompat
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.NavHostFragment
 import com.google.gson.Gson
-import com.google.gson.internal.LinkedTreeMap
 import com.hzontal.tella_vault.MyLocation
-import com.hzontal.tella_vault.VaultFile
 import org.hzontal.shared_ui.bottomsheet.BottomSheetUtils
 import org.hzontal.shared_ui.utils.DialogUtils
 import rs.readahead.washington.mobile.MyApplication
 import rs.readahead.washington.mobile.R
 import rs.readahead.washington.mobile.bus.EventObserver
 import rs.readahead.washington.mobile.bus.event.LocationPermissionRequiredEvent
-import rs.readahead.washington.mobile.data.uwazi.UwaziConstants
 import rs.readahead.washington.mobile.databinding.UwaziEntryFragmentBinding
-import rs.readahead.washington.mobile.domain.entity.collect.FormMediaFile
 import rs.readahead.washington.mobile.domain.entity.uwazi.CollectTemplate
 import rs.readahead.washington.mobile.domain.entity.uwazi.UwaziEntityInstance
 import rs.readahead.washington.mobile.domain.entity.uwazi.UwaziEntityStatus
 import rs.readahead.washington.mobile.presentation.uwazi.UwaziGeoData
-import rs.readahead.washington.mobile.presentation.uwazi.UwaziValue
-import rs.readahead.washington.mobile.presentation.uwazi.UwaziValueAttachment
 import rs.readahead.washington.mobile.util.C
 import rs.readahead.washington.mobile.views.activity.LocationMapActivity
 import rs.readahead.washington.mobile.views.base_ui.BaseBindingFragment
@@ -40,7 +34,6 @@ import rs.readahead.washington.mobile.views.fragment.uwazi.viewpager.DRAFT_LIST_
 import rs.readahead.washington.mobile.views.fragment.uwazi.viewpager.OUTBOX_LIST_PAGE_INDEX
 import rs.readahead.washington.mobile.views.fragment.uwazi.viewpager.SUBMITTED_LIST_PAGE_INDEX
 import rs.readahead.washington.mobile.views.fragment.vault.attachements.OnNavBckListener
-import timber.log.Timber
 
 
 const val COLLECT_TEMPLATE = "collect_template"
@@ -55,50 +48,16 @@ class UwaziEntryFragment :
     private val viewModel: SharedUwaziSubmissionViewModel by lazy {
         ViewModelProvider(activity).get(SharedUwaziSubmissionViewModel::class.java)
     }
+    private val uwaziParser: UwaziParser by lazy { UwaziParser(context) }
 
     private var template: CollectTemplate? = null
     private var entityInstance: UwaziEntityInstance = UwaziEntityInstance()
     private val bundle by lazy { Bundle() }
     private var screenView: ViewGroup? = null
-    private var entryPrompts = mutableListOf<UwaziEntryPrompt>()
     private lateinit var uwaziFormView: UwaziFormView
     private var hashCode: Int? = null //used to check is the answers has changed
 
-    private val uwaziTitlePrompt by lazy {
-        UwaziEntryPrompt(
-            UWAZI_TITLE,
-            "10242048",
-            UwaziConstants.UWAZI_DATATYPE_TEXT,
-            "Title",
-            true,
-            "Enter the submission title"
-        )
-    }
-
-    private val uwaziFilesPrompt by lazy {
-        UwaziEntryPrompt(
-            UWAZI_SUPPORTING_FILES,
-            "10242049",
-            UwaziConstants.UWAZI_DATATYPE_MULTIFILES,
-            getString(R.string.Uwazi_MiltiFileWidget_SupportingFiles),
-            false,
-            getString(R.string.Uwazi_MiltiFileWidget_Help)
-        )
-    }
-
     private val disposables by lazy { MyApplication.bus().createCompositeDisposable() }
-
-    private val uwaziPdfsPrompt by lazy {
-        UwaziEntryPrompt(
-            UWAZI_PRIMARY_DOCUMENTS,
-            "10242050",
-            UwaziConstants.UWAZI_DATATYPE_MULTIPDFFILES,
-            getString(R.string.Uwazi_MiltiFileWidget_PrimaryDocuments),
-            false,
-            getString(R.string.Uwazi_MiltiFileWidget_AttachMenyPdfFiles)
-        )
-    }
-
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -107,7 +66,6 @@ class UwaziEntryFragment :
             hasInitializedRootView = true
             initView()
         }
-
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
@@ -128,7 +86,13 @@ class UwaziEntryFragment :
             toolbar.backClickListener = { onBackPressed() }
             toolbar.onRightClickListener = {
                 entityInstance.status = UwaziEntityStatus.DRAFT
-                if (!getAnswersFromForm(false)) {
+                if (!uwaziParser.getAnswersFromForm(
+                        false,
+                        uwaziFormView,
+                        template,
+                        entityInstance
+                    )
+                ) {
                     uwaziFormView.setFocus(context)
                     showValidationMandatoryTitleDialog()
                 } else {
@@ -185,14 +149,14 @@ class UwaziEntryFragment :
                         showSavedDialog()
                         progress.postValue(UwaziEntityStatus.UNKNOWN)
                     }
+                    else -> {}
                 }
             })
         }
     }
 
     private fun sendEntity() {
-        //TODO REFACTOR THIS INTO A SEPARATE PARSER
-        if (!getAnswersFromForm(true)) {
+        if (!uwaziParser.getAnswersFromForm(true, uwaziFormView, template, entityInstance)) {
             uwaziFormView.setFocus(context)
             //showValidationMandatoryFieldsDialog()
             showValidationErrorsFieldsDialog()
@@ -204,178 +168,16 @@ class UwaziEntryFragment :
         }
     }
 
-    private fun getAnswersFromForm(isSend: Boolean): Boolean {
-        //TODO REFACTOR THIS INTO A SEPARATE PARSER
-        uwaziFormView.clearValidationConstraints()
-        val hashmap = mutableMapOf<String, List<Any>>()
-        val widgetMediaFiles = mutableListOf<FormMediaFile>()
-        val answers = uwaziFormView.answers
-        var validationRequired = false
-        var validationError = false
-
-        // check required fields
-        if (answers[UWAZI_TITLE] == null) {
-                uwaziFormView.setValidationConstraintText(
-                    UWAZI_TITLE,
-                    getString(R.string.Uwazi_Entity_Error_Response_Mandatory)
-                )
-            validationRequired = true
-        }
-
-        if (isSend) {
-            for (property in template?.entityRow?.properties!!) {
-                //check url validation errors
-                if (uwaziFormView.checkValidationConstraints()) {
-                    validationError = true
-                }
-
-                //check mandatory errors
-                if (property.required && (answers[property.name] == null)) {
-                    uwaziFormView.setValidationConstraintText(
-                        property.name,
-                        getString(R.string.Uwazi_Entity_Error_Response_Mandatory)
-                    )
-                    validationRequired = true
-                }
-            }
-        }
-        if (validationRequired || validationError) return false
-
-        // put answers to entity
-        for (answer in answers) {
-            if (answer.value != null) {
-                if (answer.key == UWAZI_TITLE) {
-                    entityInstance.title = (answer.value as UwaziValue).value as String
-                } else {
-                    when (answer.value) {
-                        is List<*> -> {
-                            hashmap[answer.key] = (answer.value) as List<UwaziValue>
-                        }
-                        is UwaziValueAttachment -> {
-                            hashmap[answer.key] = arrayListOf(
-                                UwaziValueAttachment(
-                                    value = (answer.value as UwaziValueAttachment).value,
-                                    attachment = uwaziFormView.filesNames.indexOf((answer.value as UwaziValueAttachment).value)
-                                )
-                            )
-                        }
-                        else -> {
-                            hashmap[answer.key] =
-                                arrayListOf(UwaziValue((answer.value as UwaziValue).value))
-                        }
-                    }
-                }
-            }
-        }
-
-        //put files in entity
-        for (answer in uwaziFormView.files) {
-            if (answer != null) {
-                widgetMediaFiles.add(answer)
-            }
-        }
-        entityInstance.metadata = hashmap
-        entityInstance.widgetMediaFiles = widgetMediaFiles
-        entityInstance.collectTemplate = template
-        entityInstance.template = template?.entityRow?.name.toString()
-        return true
-    }
-
-    private fun putAnswersToForm(instance: UwaziEntityInstance, formView: UwaziFormView) {
-        //TODO REFACTOR THIS INTO A SEPARATE PARSER
-        val files = mutableMapOf<String, FormMediaFile>()
-        for (file in instance.widgetMediaFiles) {
-            files[file.name] = file
-        }
-
-        formView.setBinaryData(UWAZI_TITLE, instance.title)
-        for (answer in instance.metadata) {
-
-            val stringVal = if ((instance.metadata[answer.key] as ArrayList).size == 1) {
-                (instance.metadata[answer.key]?.get(0) as LinkedTreeMap<String, Any>)["value"]
-            } else {
-                (instance.metadata[answer.key])
-            }
-
-            if (files.containsKey(stringVal)) {
-                formView.setBinaryData(answer.key, files[stringVal] as VaultFile)
-            } else {
-                if (stringVal != null) {
-                    formView.setBinaryData(answer.key, stringVal)
-                }
-            }
-        }
-    }
-
-    private fun fillAnswersToForm(instance: UwaziEntityInstance, formView: UwaziFormView) {
-        //TODO REFACTOR THIS INTO A SEPARATE PARSER
-        val files = mutableMapOf<String, FormMediaFile>()
-        for (file in instance.widgetMediaFiles) {
-            files[file.name] = file
-        }
-
-        formView.setBinaryData(UWAZI_TITLE, instance.title)
-
-        for (answer in instance.metadata) {
-            if ((answer.value as List<*>).size > 1) {
-                formView.setBinaryData(answer.key, answer.value)
-            } else {
-                val uwaziValue: UwaziValue = answer.value[0] as UwaziValue
-                val stringVal = uwaziValue.value
-                if (files.containsKey(stringVal)) {
-                    formView.setBinaryData(answer.key, files[stringVal] as VaultFile)
-                } else {
-                    formView.setBinaryData(answer.key, stringVal)
-                }
-            }
-        }
-    }
-
     private fun parseUwaziInstance() {
-        prepareFormView()
-        putAnswersToForm(entityInstance, uwaziFormView)
+        uwaziFormView = uwaziParser.prepareFormView(template)
+        screenView?.addView(uwaziFormView)
+        uwaziParser.putAnswersToForm(entityInstance, uwaziFormView)
     }
 
     private fun parseUwaziTemplate() {
-        prepareFormView()
-        fillAnswersToForm(entityInstance, uwaziFormView)
-    }
-
-    private fun prepareFormView() {
-        //TODO REFACTOR THIS INTO A SEPARATE PARSER
-        entryPrompts.clear()
-
-        //TODO Handle this special common props smarter
-        entryPrompts.add(uwaziPdfsPrompt)
-        entryPrompts.add(uwaziFilesPrompt)
-
-        if (template?.entityRow?.commonProperties?.get(0)?.translatedLabel?.length!! > 0) {
-            uwaziTitlePrompt.question =
-                template?.entityRow?.commonProperties?.get(0)?.translatedLabel
-        }
-        entryPrompts.add(uwaziTitlePrompt)
-
-        for (property in template?.entityRow?.properties!!) {
-            val entryPrompt = UwaziEntryPrompt(
-                property.name,
-                property.id,
-                property.type,
-                property.translatedLabel,
-                property.required,
-                property.translatedLabel
-            )
-            if (property.values != null) {
-                entryPrompt.selectValues = property.values
-            }
-            entryPrompts.add(entryPrompt)
-        }
-
-        val arr: Array<UwaziEntryPrompt?> = arrayOfNulls(entryPrompts.size)
-        arr.indices.forEach { i ->
-            arr[i] = entryPrompts[i]
-        }
-        uwaziFormView = UwaziFormView(requireContext(), arr)
+        uwaziFormView = uwaziParser.prepareFormView(template)
         screenView?.addView(uwaziFormView)
+        uwaziParser.fillAnswersToForm(entityInstance, uwaziFormView)
     }
 
     private fun putVaultFileInForm(vaultFile: String) {
@@ -442,7 +244,7 @@ class UwaziEntryFragment :
             object : EventObserver<LocationPermissionRequiredEvent?>() {
                 override fun onNext(event: LocationPermissionRequiredEvent) {
                     if (!hasGpsPermissions(requireContext())) {
-                        activity.maybeChangeTemporaryTimeout{
+                        activity.maybeChangeTemporaryTimeout {
                             requestGpsPermissions(C.GPS_PROVIDER)
                         }
                     }
@@ -452,7 +254,9 @@ class UwaziEntryFragment :
 
     override fun onBackPressed(): Boolean {
         // The save draft dialog should be shown if the form could be saved and if the answers have changed
-        if (getAnswersFromForm(false) && hashCode != uwaziFormView.answers.hashCode()) {
+        if (uwaziParser.getAnswersFromForm(false, uwaziFormView, template, entityInstance)
+            && hashCode != uwaziFormView.answers.hashCode()
+        ) {
             BottomSheetUtils.showStandardSheet(
                 activity.supportFragmentManager,
                 activity.getString(R.string.Uwazi_Dialog_Draft_Title),
